1번회원을 관리자로 지정, 게시물 수정/삭제시 권한체크

+++ src/main/java/com/sbs/untact/controller/UsrArticleController.java
@@ -59,7 +59,7 @@ public class UsrArticleController {
 	public ResultData doAdd(@RequestParam Map<String, Object> param, HttpSession session) {
 		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
 		if (loginedMemberId == 0) {
-			return new ResultData("F-2", "고그인 후 이용해 주세요.");
+			return new ResultData("F-2", "로그인 후 이용해 주세요.");
 		}
 		
 		if (param.get("title") == null) {
@@ -69,13 +69,18 @@ public class UsrArticleController {
 		if (param.get("body") == null) {
 			return new ResultData("F-1", "body를 입력해주세요.");
 		}
-
+		param.put("memberId", loginedMemberId);
 		return articleService.addArticle(param);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if (loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해 주세요.");
+		}
+		
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -85,13 +90,22 @@ public class UsrArticleController {
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
+		ResultData getActorCanDeleteRd= articleService.getActorCanDeleteRd(article, loginedMemberId);
+		if(getActorCanDeleteRd.isFail()) {
+			return getActorCanDeleteRd;
+		}
 
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if (loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해 주세요.");
+		}
+		
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -108,6 +122,10 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+		ResultData getActorCanModifyRd= articleService.getActorCanModifyRd(article, loginedMemberId);
+		if(getActorCanModifyRd.isFail()) {
+			return getActorCanModifyRd;
 		}
 
 		return articleService.modifyArticle(id, title, body);
